+++ src/components/Elements/Modal/ModalFooter.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 /**
- * The ModalFooter compoent manage the  footer of the modal component.
+ * The ModalFooter component manage the  footer of the modal component.
  */
 const ModalFooter = ({ children }) => (<footer className="bi-modal-footer">{children}</footer>);